better organization of plot_utils
