+++ simulator/src/scenarios/plot_utils.py
@@ -1377,170 +1377,7 @@ def generate_individual_curves_plots(data: Dict[str, Any], param_name: str, resu
         import traceback
         traceback.print_exc()
 
-def calculate_delta_from_counts(count_data: List[Tuple[int, int]]) -> List[Tuple[int, int]]:
-    """
-    Calculate delta (change) between consecutive transaction counts.
-    
-    For each index i, subtract the value at index i-1 from the value at index i.
-    The first entry (i=0) will have delta = 0.
-    
-    Args:
-        count_data: List of tuples (height, count) representing transaction counts over time
-        
-    Returns:
-        List of tuples (height, delta) representing the change in counts between consecutive time steps
-    """
-    if not count_data:
-        return []
-    
-    delta_data = []
-    
-    # First entry has delta = 0
-    first_height, first_count = count_data[0]
-    delta_data.append((first_height, 0))
-    
-    # Calculate deltas for subsequent entries
-    for i in range(1, len(count_data)):
-        current_height, current_count = count_data[i]
-        previous_count = count_data[i-1][1]
-        delta = current_count - previous_count
-        delta_data.append((current_height, delta))
-    
-    return delta_data
 
-def plot_transactions_delta_overlay(
-    data: Dict[str, Any],
-    param_name: str,
-    transaction_type: str,
-    results_dir: str,
-    sweep_type: str
-) -> None:
-    """Plot transaction delta overlay for a specific transaction type"""
-    try:
-        individual_results = data['individual_results']
-        
-        if not individual_results:
-            print(f"Warning: No individual results found, skipping {transaction_type} transactions delta plot")
-            return
-        
-        # Create figure
-        plt.figure(figsize=(10, 6))
-        
-        # Create color gradient
-        colors = create_color_gradient(len(individual_results))
-        
-        # Track maximum height for xlim
-        max_height = 0
-        
-        # Plot each simulation's chain 1 transactions
-        for i, result in enumerate(individual_results):
-            param_value = extract_parameter_value(result, param_name)
-            
-            # For main transaction types (pending, success, failure), calculate as CAT + regular
-            if transaction_type in ['pending', 'success', 'failure']:
-                cat_data = result.get(f'chain_1_cat_{transaction_type}', [])
-                regular_data = result.get(f'chain_1_regular_{transaction_type}', [])
-                
-                # Create a combined dataset by summing CAT and regular at each height
-                combined_data = {}
-                
-                # Add CAT data
-                for height, count in cat_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Add regular data
-                for height, count in regular_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Convert back to sorted list of tuples
-                chain_data = sorted(combined_data.items())
-            else:
-                # For CAT and regular specific types, use the data directly
-                chain_data = result[f'chain_1_{transaction_type}']
-            
-            if not chain_data:
-                continue
-            
-            # Calculate deltas from the count data
-            delta_data = calculate_delta_from_counts(chain_data)
-            
-            if not delta_data:
-                continue
-            
-            # Trim the last 10% of data to avoid edge effects
-            delta_data = trim_time_series_data(delta_data, 0.1)
-            
-            if not delta_data:
-                continue
-                
-            # Extract data - delta_data is a list of tuples (height, delta)
-            heights = [entry[0] for entry in delta_data]
-            deltas = [entry[1] for entry in delta_data]
-            
-            # Update maximum height
-            if heights:
-                max_height = max(max_height, max(heights))
-            
-            # Plot with color based on parameter
-            label = create_parameter_label(param_name, param_value)
-            plt.plot(heights, deltas, color=colors[i], alpha=0.7, 
-                    label=label, linewidth=1.5)
-        
-        # Set x-axis limits before finalizing the plot
-        plt.xlim(left=0, right=max_height)
-        
-        # Create title and filename based on transaction type
-        if transaction_type in ['pending', 'success', 'failure']:
-            # Combined totals
-            title = f'SumTypes {transaction_type.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}_sumTypes.png'
-        elif transaction_type.startswith('cat_'):
-            # CAT transactions
-            status = transaction_type.replace('cat_', '')
-            if status == 'pending_resolving':
-                title = f'CAT Pending Resolving Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_resolving.png'
-            elif status == 'pending_postponed':
-                title = f'CAT Pending Postponed Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_postponed.png'
-            else:
-                title = f'CAT {status.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_{status}_cat.png'
-        elif transaction_type.startswith('regular_'):
-            # Regular transactions
-            status = transaction_type.replace('regular_', '')
-            title = f'Regular {status.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{status}_regular.png'
-        else:
-            # Fallback
-            title = f'{transaction_type.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}.png'
-        
-        plt.title(title)
-        plt.xlabel('Block Height')
-        plt.ylabel(f'Delta in {transaction_type.title()} Transactions')
-        plt.grid(True, alpha=0.3)
-        plt.legend(loc="upper right")
-        plt.tight_layout()
-        
-        # Create tx_count_delta directory and save plot
-        tx_count_delta_dir = f'{results_dir}/figs/tx_count_delta'
-        os.makedirs(tx_count_delta_dir, exist_ok=True)
-        plt.savefig(f'{tx_count_delta_dir}/{filename}', 
-                   dpi=300, bbox_inches='tight')
-        plt.close()
-        
-    except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
-        print(f"Warning: Error processing {transaction_type} transactions delta data: {e}")
-        return
-
-def plot_cat_pending_resolving_delta(data: Dict[str, Any], param_name: str, results_dir: str, sweep_type: str) -> None:
-    """Plot CAT pending resolving transaction deltas overlay"""
-    plot_transactions_delta_overlay(data, param_name, 'cat_pending_resolving', results_dir, sweep_type)
-
-def plot_cat_pending_postponed_delta(data: Dict[str, Any], param_name: str, results_dir: str, sweep_type: str) -> None:
-    """Plot CAT pending postponed transaction deltas overlay"""
-    plot_transactions_delta_overlay(data, param_name, 'cat_pending_postponed', results_dir, sweep_type)
 
 def run_sweep_plots(sweep_name: str, param_name: str, sweep_type: str) -> None:
     """
@@ -1603,813 +1440,9 @@ def load_plot_config(results_dir: str) -> Dict[str, Any]:
         'cutoff': plot_config['cutoff']
     }
 
-def apply_moving_average(data: List[Tuple[int, int]], window_size: int) -> List[Tuple[int, int]]:
-    """
-    Apply moving average to time series data.
-    
-    Args:
-        data: List of tuples (height, value) representing time series data
-        window_size: Number of consecutive points to average
-        
-    Returns:
-        List of tuples (height, averaged_value) with moving average applied
-    """
-    if len(data) < window_size:
-        return data
-    
-    averaged_data = []
-    
-    for i in range(len(data)):
-        # Calculate start and end indices for the window
-        start_idx = max(0, i - window_size + 1)
-        end_idx = i + 1
-        
-        # Extract values in the window
-        window_values = [data[j][1] for j in range(start_idx, end_idx)]
-        
-        # Calculate average
-        avg_value = sum(window_values) / len(window_values)
-        
-        # Keep the original height, use averaged value
-        averaged_data.append((data[i][0], avg_value))
-    
-    return averaged_data
-
-def plot_transactions_overlay_with_moving_average(
-    data: Dict[str, Any],
-    param_name: str,
-    transaction_type: str,
-    results_dir: str,
-    sweep_type: str,
-    plot_config: Dict[str, Any]
-) -> None:
-    """Plot transaction overlay with optional moving average"""
-    try:
-        individual_results = data['individual_results']
-        
-        if not individual_results:
-            print(f"Warning: No individual results found, skipping {transaction_type} transactions plot")
-            return
-        
-        # Create figure
-        plt.figure(figsize=(10, 6))
-        
-        # Create color gradient
-        colors = create_color_gradient(len(individual_results))
-        
-        # Track maximum height for xlim
-        max_height = 0
-        
-        # Plot each simulation's chain 1 transactions
-        for i, result in enumerate(individual_results):
-            param_value = extract_parameter_value(result, param_name)
-            
-            # For main transaction types (pending, success, failure), calculate as CAT + regular
-            if transaction_type in ['pending', 'success', 'failure']:
-                cat_data = result.get(f'chain_1_cat_{transaction_type}', [])
-                regular_data = result.get(f'chain_1_regular_{transaction_type}', [])
-                
-                # Create a combined dataset by summing CAT and regular at each height
-                combined_data = {}
-                
-                # Add CAT data
-                for height, count in cat_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Add regular data
-                for height, count in regular_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Convert back to sorted list of tuples
-                chain_data = sorted(combined_data.items())
-            else:
-                # For CAT and regular specific types, use the data directly
-                chain_data = result[f'chain_1_{transaction_type}']
-            
-            if not chain_data:
-                continue
-            
-            # Apply moving average if enabled
-            if plot_config.get('plot_moving_average', False):
-                window_size = plot_config.get('range_moving_average', 10)
-                chain_data = apply_moving_average(chain_data, window_size)
-            
-            # Trim the last 10% of data to avoid edge effects
-            chain_data = trim_time_series_data(chain_data, 0.1)
-            
-            if not chain_data:
-                continue
-                
-            # Extract data - chain_data is a list of tuples (height, count)
-            heights = [entry[0] for entry in chain_data]
-            counts = [entry[1] for entry in chain_data]
-            
-            # Update maximum height
-            if heights:
-                max_height = max(max_height, max(heights))
-            
-            # Plot with color based on parameter
-            label = create_parameter_label(param_name, param_value)
-            plt.plot(heights, counts, color=colors[i], alpha=0.7, 
-                    label=label, linewidth=1.5)
-        
-        # Set x-axis limits before finalizing the plot
-        plt.xlim(left=0, right=max_height)
-        
-        # Create title and filename based on transaction type
-        if transaction_type in ['pending', 'success', 'failure']:
-            # Combined totals
-            title = f'SumTypes {transaction_type.title()} Transactions by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}_sumTypes.png'
-        elif transaction_type.startswith('cat_'):
-            # CAT transactions
-            status = transaction_type.replace('cat_', '')
-            if status == 'pending_resolving':
-                title = f'CAT Pending Resolving Transactions by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_resolving.png'
-            elif status == 'pending_postponed':
-                title = f'CAT Pending Postponed Transactions by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_postponed.png'
-            else:
-                title = f'CAT {status.title()} Transactions by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_{status}_cat.png'
-        elif transaction_type.startswith('regular_'):
-            # Regular transactions
-            status = transaction_type.replace('regular_', '')
-            title = f'Regular {status.title()} Transactions by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{status}_regular.png'
-        else:
-            # Fallback
-            title = f'{transaction_type.title()} Transactions by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}.png'
-        
-        # Add moving average indicator to title if enabled
-        if plot_config.get('plot_moving_average', False):
-            window_size = plot_config.get('range_moving_average', 10)
-            title += f" (Moving Average, Window={window_size})"
-        
-        plt.title(title)
-        plt.xlabel('Block Height')
-        plt.ylabel(f'Number of {transaction_type.title()} Transactions')
-        plt.grid(True, alpha=0.3)
-        plt.legend(loc="upper right")
-        plt.tight_layout()
-        
-        # Determine directory based on whether moving average is applied
-        if plot_config.get('plot_moving_average', False):
-            base_dir = f'{results_dir}/figs/tx_count/moving_average'
-        else:
-            base_dir = f'{results_dir}/figs/tx_count'
-        
-        # Create directory and save plot
-        os.makedirs(base_dir, exist_ok=True)
-        plt.savefig(f'{base_dir}/{filename}', 
-                   dpi=300, bbox_inches='tight')
-        plt.close()
-        
-    except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
-        print(f"Warning: Error processing {transaction_type} transactions data: {e}")
-        return
-
-def plot_transactions_delta_overlay_with_moving_average(
-    data: Dict[str, Any],
-    param_name: str,
-    transaction_type: str,
-    results_dir: str,
-    sweep_type: str,
-    plot_config: Dict[str, Any]
-) -> None:
-    """Plot transaction delta overlay with optional moving average"""
-    try:
-        individual_results = data['individual_results']
-        
-        if not individual_results:
-            print(f"Warning: No individual results found, skipping {transaction_type} transactions delta plot")
-            return
-        
-        # Create figure
-        plt.figure(figsize=(10, 6))
-        
-        # Create color gradient
-        colors = create_color_gradient(len(individual_results))
-        
-        # Track maximum height for xlim
-        max_height = 0
-        
-        # Plot each simulation's chain 1 transactions
-        for i, result in enumerate(individual_results):
-            param_value = extract_parameter_value(result, param_name)
-            
-            # For main transaction types (pending, success, failure), calculate as CAT + regular
-            if transaction_type in ['pending', 'success', 'failure']:
-                cat_data = result.get(f'chain_1_cat_{transaction_type}', [])
-                regular_data = result.get(f'chain_1_regular_{transaction_type}', [])
-                
-                # Create a combined dataset by summing CAT and regular at each height
-                combined_data = {}
-                
-                # Add CAT data
-                for height, count in cat_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Add regular data
-                for height, count in regular_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Convert back to sorted list of tuples
-                chain_data = sorted(combined_data.items())
-            else:
-                # For CAT and regular specific types, use the data directly
-                chain_data = result[f'chain_1_{transaction_type}']
-            
-            if not chain_data:
-                continue
-            
-            # Calculate deltas from the count data
-            delta_data = calculate_delta_from_counts(chain_data)
-            
-            if not delta_data:
-                continue
-            
-            # Apply moving average if enabled
-            if plot_config.get('plot_moving_average', False):
-                window_size = plot_config.get('range_moving_average', 10)
-                delta_data = apply_moving_average(delta_data, window_size)
-            
-            # Trim the last 10% of data to avoid edge effects
-            delta_data = trim_time_series_data(delta_data, 0.1)
-            
-            if not delta_data:
-                continue
-                
-            # Extract data - delta_data is a list of tuples (height, delta)
-            heights = [entry[0] for entry in delta_data]
-            deltas = [entry[1] for entry in delta_data]
-            
-            # Update maximum height
-            if heights:
-                max_height = max(max_height, max(heights))
-            
-            # Plot with color based on parameter
-            label = create_parameter_label(param_name, param_value)
-            plt.plot(heights, deltas, color=colors[i], alpha=0.7, 
-                    label=label, linewidth=1.5)
-        
-        # Set x-axis limits before finalizing the plot
-        plt.xlim(left=0, right=max_height)
-        
-        # Create title and filename based on transaction type
-        if transaction_type in ['pending', 'success', 'failure']:
-            # Combined totals
-            title = f'SumTypes {transaction_type.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}_sumTypes.png'
-        elif transaction_type.startswith('cat_'):
-            # CAT transactions
-            status = transaction_type.replace('cat_', '')
-            if status == 'pending_resolving':
-                title = f'CAT Pending Resolving Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_resolving.png'
-            elif status == 'pending_postponed':
-                title = f'CAT Pending Postponed Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_postponed.png'
-            else:
-                title = f'CAT {status.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_{status}_cat.png'
-        elif transaction_type.startswith('regular_'):
-            # Regular transactions
-            status = transaction_type.replace('regular_', '')
-            title = f'Regular {status.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{status}_regular.png'
-        else:
-            # Fallback
-            title = f'{transaction_type.title()} Transaction Deltas by Height (Chain 1) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}.png'
-        
-        # Add moving average indicator to title if enabled
-        if plot_config.get('plot_moving_average', False):
-            window_size = plot_config.get('range_moving_average', 10)
-            title += f" (Moving Average, Window={window_size})"
-        
-        plt.title(title)
-        plt.xlabel('Block Height')
-        plt.ylabel(f'Delta in {transaction_type.title()} Transactions')
-        plt.grid(True, alpha=0.3)
-        plt.legend(loc="upper right")
-        plt.tight_layout()
-        
-        # Determine directory based on whether moving average is applied
-        if plot_config.get('plot_moving_average', False):
-            base_dir = f'{results_dir}/figs/tx_count_delta/moving_average'
-        else:
-            base_dir = f'{results_dir}/figs/tx_count_delta'
-        
-        # Create directory and save plot
-        os.makedirs(base_dir, exist_ok=True)
-        plt.savefig(f'{base_dir}/{filename}', 
-                   dpi=300, bbox_inches='tight')
-        plt.close()
-        
-    except (FileNotFoundError, json.JSONDecodeError, KeyError) as e:
-        print(f"Warning: Error processing {transaction_type} transactions delta data: {e}")
-        return
-
-def apply_cutoff_to_data(data: List[Tuple[int, int]], cutoff_height: int, transaction_type: str) -> List[Tuple[int, int]]:
-    """
-    Apply cutoff to time series data by removing data before cutoff_height and subtracting offset values.
-    
-    Args:
-        data: List of tuples (height, value) representing time series data
-        cutoff_height: Height at which to apply the cutoff
-        transaction_type: Type of transaction (success, failure, pending, etc.)
-        
-    Returns:
-        List of tuples (height, adjusted_value) with cutoff applied
-    """
-    if not data:
-        return []
-    
-    # Find the offset value at the cutoff height
-    offset_value = 0
-    for height, value in data:
-        if height >= cutoff_height:
-            offset_value = value
-            break
-    
-    # Apply cutoff: remove data before cutoff_height
-    cutoff_data = []
-    for height, value in data:
-        if height >= cutoff_height:
-            # Only subtract offset for success and failure transactions (including cat_ and regular_ prefixes)
-            if 'success' in transaction_type or 'failure' in transaction_type:
-                adjusted_value = value - offset_value
-            else:
-                # For pending and other types, just use the original value
-                adjusted_value = value
-            cutoff_data.append((height, adjusted_value))
-    
-    return cutoff_data
-
-def plot_transactions_cutoff_overlay(
-    data: Dict[str, Any],
-    param_name: str,
-    transaction_type: str,
-    results_dir: str,
-    sweep_type: str,
-    plot_config: Dict[str, Any]
-) -> None:
-    """
-    Plot transaction overlay with cutoff applied.
-    
-    This function:
-    1. Removes all data before cutoff_height (sim_total_block_number * cutoff)
-    2. For success, failure, and pending postponed transactions, records offset values at cutoff
-    3. Subtracts these offset values from the remainder of the data vector
-    4. Everything else stays the same
-    """
-    try:
-        individual_results = data['individual_results']
-        
-        if not individual_results:
-            print(f"Warning: No individual results found, skipping {transaction_type} cutoff plot")
-            return
-        
-        # Get cutoff configuration
-        cutoff_percentage = plot_config.get('cutoff', 0.5)
-        
-        # Create figure
-        plt.figure(figsize=(10, 6))
-        
-        # Create color gradient
-        colors = create_color_gradient(len(individual_results))
-        
-        # Track maximum height for xlim
-        max_height = 0
-        
-        # Plot each simulation's chain 1 transactions
-        for i, result in enumerate(individual_results):
-            param_value = extract_parameter_value(result, param_name)
-            
-            # For main transaction types (pending, success, failure), calculate as CAT + regular
-            if transaction_type in ['pending', 'success', 'failure']:
-                cat_data = result.get(f'chain_1_cat_{transaction_type}', [])
-                regular_data = result.get(f'chain_1_regular_{transaction_type}', [])
-                
-                # Create a combined dataset by summing CAT and regular at each height
-                combined_data = {}
-                
-                # Add CAT data
-                for height, count in cat_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Add regular data
-                for height, count in regular_data:
-                    combined_data[height] = combined_data.get(height, 0) + count
-                
-                # Convert back to sorted list of tuples
-                chain_data = sorted(combined_data.items())
-            else:
-                # For CAT and regular specific types, use the data directly
-                chain_data = result[f'chain_1_{transaction_type}']
-            
-            if not chain_data:
-                continue
-            
-            # Calculate cutoff height based on maximum height in data
-            max_data_height = max(chain_data, key=lambda x: x[0])[0]
-            cutoff_height = int(max_data_height * cutoff_percentage)
-            
-            # Apply cutoff to the data
-            cutoff_data = apply_cutoff_to_data(chain_data, cutoff_height, transaction_type)
-            
-            if not cutoff_data:
-                continue
-            
-            # Trim the last 10% of data to avoid edge effects
-            cutoff_data = trim_time_series_data(cutoff_data, 0.1)
-            
-            if not cutoff_data:
-                continue
-                
-            # Extract data - cutoff_data is a list of tuples (height, adjusted_value)
-            heights = [entry[0] for entry in cutoff_data]
-            values = [entry[1] for entry in cutoff_data]
-            
-            # Update maximum height
-            if heights:
-                max_height = max(max_height, max(heights))
-            
-            # Plot with color based on parameter
-            label = create_parameter_label(param_name, param_value)
-            plt.plot(heights, values, color=colors[i], alpha=0.7, 
-                    label=label, linewidth=1.5)
-        
-        # Set x-axis limits before finalizing the plot
-        plt.xlim(left=0, right=max_height)
-        
-        # Create title and filename based on transaction type
-        if transaction_type in ['pending', 'success', 'failure']:
-            # Combined totals
-            title = f'SumTypes {transaction_type.title()} Transaction Counts (Cutoff Applied) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}_sumTypes_cutoff.png'
-        elif transaction_type.startswith('cat_'):
-            # CAT transactions
-            status = transaction_type.replace('cat_', '')
-            if status == 'pending_resolving':
-                title = f'CAT Pending Resolving Transaction Counts (Cutoff Applied) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_resolving_cutoff.png'
-            elif status == 'pending_postponed':
-                title = f'CAT Pending Postponed Transaction Counts (Cutoff Applied) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_pending_cat_postponed_cutoff.png'
-            else:
-                title = f'CAT {status.title()} Transaction Counts (Cutoff Applied) - {create_sweep_title(param_name, sweep_type)}'
-                filename = f'tx_{status}_cat_cutoff.png'
-        elif transaction_type.startswith('regular_'):
-            # Regular transactions
-            status = transaction_type.replace('regular_', '')
-            title = f'Regular {status.title()} Transaction Counts (Cutoff Applied) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{status}_regular_cutoff.png'
-        else:
-            # Fallback
-            title = f'{transaction_type.title()} Transaction Counts (Cutoff Applied) - {create_sweep_title(param_name, sweep_type)}'
-            filename = f'tx_{transaction_type}_cutoff.png'
-        
-        plt.title(title, fontsize=14)
-        plt.xlabel('Block Height', fontsize=12)
-        plt.ylabel('Transaction Count (Adjusted)', fontsize=12)
-        plt.grid(True, alpha=0.3)
-        plt.legend(loc="upper right", fontsize=10)
-        plt.tight_layout()
-        
-        # Create tx_count_cutoff directory and save plot
-        tx_count_cutoff_dir = f'{results_dir}/figs/tx_count_cutoff'
-        os.makedirs(tx_count_cutoff_dir, exist_ok=True)
-        plt.savefig(f'{tx_count_cutoff_dir}/{filename}', 
-                   dpi=300, bbox_inches='tight')
-        plt.close()
-        
-        print(f"Generated cutoff plot: {filename}")
-        
-    except Exception as e:
-        print(f"Error generating cutoff plot for {transaction_type}: {e}")
-        import traceback
-        traceback.print_exc()
 
 
 
-def apply_cutoff_to_percentage_data(data: Dict[str, Any], plot_config: Dict[str, Any]) -> Dict[str, Any]:
-    """
-    Apply cutoff to all transaction data in the results before percentage calculation.
-    
-    Args:
-        data: The original sweep data
-        plot_config: Plot configuration containing cutoff settings
-        
-    Returns:
-        Modified data with cutoff applied to all transaction data
-    """
-    cutoff_percentage = plot_config.get('cutoff', 0.5)
-    modified_data = {'individual_results': []}
-    
-    for result in data['individual_results']:
-        modified_result = result.copy()
-        
-        # Apply cutoff to all transaction data
-        transaction_keys = [
-            'chain_1_cat_success', 'chain_1_cat_failure', 'chain_1_cat_pending',
-            'chain_1_regular_success', 'chain_1_regular_failure', 'chain_1_regular_pending',
-            'chain_1_cat_pending_resolving', 'chain_1_cat_pending_postponed'
-        ]
-        
-        for key in transaction_keys:
-            if key in modified_result:
-                tx_data = modified_result[key]
-                if tx_data:
-                    # Calculate cutoff height based on maximum height in data
-                    max_data_height = max(tx_data, key=lambda x: x[0])[0]
-                    cutoff_height = int(max_data_height * cutoff_percentage)
-                    
-                    # Apply cutoff to the data
-                    cutoff_data = apply_cutoff_to_data(tx_data, cutoff_height, key)
-                    modified_result[key] = cutoff_data
-        
-        modified_data['individual_results'].append(modified_result)
-    
-    return modified_data
-
-def plot_transaction_percentage_cutoff(
-    data: Dict[str, Any], 
-    param_name: str, 
-    results_dir: str, 
-    sweep_type: str, 
-    transaction_type: str, 
-    percentage_type: str,
-    plot_config: Dict[str, Any]
-) -> None:
-    """
-    Plot transaction percentage with cutoff applied.
-    
-    This function applies cutoff to the data and creates percentage plots
-    saved to the tx_count_cutoff directory.
-    """
-    # Apply cutoff to the data
-    cutoff_data = apply_cutoff_to_percentage_data(data, plot_config)
-    
-    # Import the percentage plotting function
-    from plot_utils_percentage import plot_transaction_percentage
-    
-    # Create a modified version that saves to the correct directory
-    def plot_transaction_percentage_cutoff_save(cutoff_data, param_name, results_dir, sweep_type, transaction_type, percentage_type):
-        """Modified version of plot_transaction_percentage that saves to tx_count_cutoff directory"""
-        try:
-            # Extract parameter values and results
-            param_values = []
-            results = []
-            
-            for result in cutoff_data['individual_results']:
-                param_value = extract_parameter_value(result, param_name)
-                param_values.append(param_value)
-                results.append(result)
-            
-            # Create color gradient
-            from plot_utils_percentage import create_color_gradient
-            colors = create_color_gradient(len(param_values))
-            
-            # Create the plot
-            plt.figure(figsize=(10, 6))
-            
-            # Track maximum height for xlim
-            max_height = 0
-            
-            # Plot each parameter value
-            for i, (param_value, result) in enumerate(zip(param_values, results)):
-                # Extract transaction data using the same structure as regular plots
-                if transaction_type == 'cat':
-                    # For CAT transactions, get the specific status data
-                    if percentage_type == 'success':
-                        tx_data = result.get('chain_1_cat_success', [])
-                    elif percentage_type == 'failure':
-                        tx_data = result.get('chain_1_cat_failure', [])
-                    else:  # pending
-                        tx_data = result.get('chain_1_cat_pending', [])
-                elif transaction_type == 'regular':
-                    # For regular transactions, get the specific status data
-                    if percentage_type == 'success':
-                        tx_data = result.get('chain_1_regular_success', [])
-                    elif percentage_type == 'failure':
-                        tx_data = result.get('chain_1_regular_failure', [])
-                    else:  # pending
-                        tx_data = result.get('chain_1_regular_pending', [])
-                else:  # sumtypes
-                    # For sumtypes, combine CAT and regular data
-                    if percentage_type == 'success':
-                        cat_data = result.get('chain_1_cat_success', [])
-                        regular_data = result.get('chain_1_regular_success', [])
-                    elif percentage_type == 'failure':
-                        cat_data = result.get('chain_1_cat_failure', [])
-                        regular_data = result.get('chain_1_regular_failure', [])
-                    else:  # pending
-                        cat_data = result.get('chain_1_cat_pending', [])
-                        regular_data = result.get('chain_1_regular_pending', [])
-                    
-                    # Create a combined dataset by summing CAT and regular at each height
-                    combined_data = {}
-                    
-                    # Add CAT data
-                    for height, count in cat_data:
-                        combined_data[height] = combined_data.get(height, 0) + count
-                    
-                    # Add regular data
-                    for height, count in regular_data:
-                        combined_data[height] = combined_data.get(height, 0) + count
-                    
-                    # Convert back to sorted list of tuples
-                    tx_data = sorted(combined_data.items())
-                
-                if not tx_data:
-                    continue
-                
-                # Extract heights and transaction counts
-                heights = []
-                percentages = []
-                
-                # Convert tx_data to list of tuples if it's not already
-                if isinstance(tx_data, list) and tx_data and isinstance(tx_data[0], dict):
-                    # If it's a list of dictionaries, convert to list of tuples
-                    tx_data = [(entry.get('height', 0), entry.get('count', 0)) for entry in tx_data]
-                
-                # Process each block - tx_data is now a list of tuples (height, count)
-                for height, count in tx_data:
-                    heights.append(height)
-                    
-                    # Calculate percentage using counts at this specific height (not cumulative)
-                    if transaction_type == 'cat':
-                        # For CAT transactions, get all status data
-                        cat_success_data = result.get('chain_1_cat_success', [])
-                        cat_failure_data = result.get('chain_1_cat_failure', [])
-                        cat_pending_data = result.get('chain_1_cat_pending', [])
-                        
-                        # Convert to height->count mapping
-                        cat_success_by_height = {entry[0]: entry[1] for entry in cat_success_data}
-                        cat_failure_by_height = {entry[0]: entry[1] for entry in cat_failure_data}
-                        cat_pending_by_height = {entry[0]: entry[1] for entry in cat_pending_data}
-                        
-                        # Get counts at this specific height
-                        success_at_height = cat_success_by_height.get(height, 0)
-                        failure_at_height = cat_failure_by_height.get(height, 0)
-                        pending_at_height = cat_pending_by_height.get(height, 0)
-                        
-                    elif transaction_type == 'regular':
-                        # For regular transactions, get all status data
-                        regular_success_data = result.get('chain_1_regular_success', [])
-                        regular_failure_data = result.get('chain_1_regular_failure', [])
-                        regular_pending_data = result.get('chain_1_regular_pending', [])
-                        
-                        # Convert to height->count mapping
-                        regular_success_by_height = {entry[0]: entry[1] for entry in regular_success_data}
-                        regular_failure_by_height = {entry[0]: entry[1] for entry in regular_failure_data}
-                        regular_pending_by_height = {entry[0]: entry[1] for entry in regular_pending_data}
-                        
-                        # Get counts at this specific height
-                        success_at_height = regular_success_by_height.get(height, 0)
-                        failure_at_height = regular_failure_by_height.get(height, 0)
-                        pending_at_height = regular_pending_by_height.get(height, 0)
-                        
-                    elif transaction_type in ['cat_pending_resolving', 'cat_pending_postponed']:
-                        # For CAT pending resolving/postponed, use (resolving + postponed) as denominator
-                        cat_pending_resolving_data = result.get('chain_1_cat_pending_resolving', [])
-                        cat_pending_postponed_data = result.get('chain_1_cat_pending_postponed', [])
-                        
-                        # Convert to height->count mapping
-                        cat_pending_resolving_by_height = {entry[0]: entry[1] for entry in cat_pending_resolving_data}
-                        cat_pending_postponed_by_height = {entry[0]: entry[1] for entry in cat_pending_postponed_data}
-                        
-                        # Get counts at this specific height
-                        resolving_at_height = cat_pending_resolving_by_height.get(height, 0)
-                        postponed_at_height = cat_pending_postponed_by_height.get(height, 0)
-                        
-                        # Calculate percentage of resolving/postponed vs total pending (resolving + postponed)
-                        total_pending = resolving_at_height + postponed_at_height
-                        if total_pending > 0:
-                            if transaction_type == 'cat_pending_resolving':
-                                percentage = (resolving_at_height / total_pending) * 100
-                            else:  # cat_pending_postponed
-                                percentage = (postponed_at_height / total_pending) * 100
-                        else:
-                            percentage = 0
-                        
-                    else:  # sumtypes
-                        # For sumtypes, combine CAT and regular data
-                        cat_success_data = result.get('chain_1_cat_success', [])
-                        cat_failure_data = result.get('chain_1_cat_failure', [])
-                        cat_pending_data = result.get('chain_1_cat_pending', [])
-                        regular_success_data = result.get('chain_1_regular_success', [])
-                        regular_failure_data = result.get('chain_1_regular_failure', [])
-                        regular_pending_data = result.get('chain_1_regular_pending', [])
-                        
-                        # Convert to height->count mapping
-                        cat_success_by_height = {entry[0]: entry[1] for entry in cat_success_data}
-                        cat_failure_by_height = {entry[0]: entry[1] for entry in cat_failure_data}
-                        cat_pending_by_height = {entry[0]: entry[1] for entry in cat_pending_data}
-                        regular_success_by_height = {entry[0]: entry[1] for entry in regular_success_data}
-                        regular_failure_by_height = {entry[0]: entry[1] for entry in regular_failure_data}
-                        regular_pending_by_height = {entry[0]: entry[1] for entry in regular_pending_data}
-                        
-                        # Get combined counts at this specific height (CAT + regular)
-                        success_at_height = cat_success_by_height.get(height, 0) + regular_success_by_height.get(height, 0)
-                        failure_at_height = cat_failure_by_height.get(height, 0) + regular_failure_by_height.get(height, 0)
-                        pending_at_height = cat_pending_by_height.get(height, 0) + regular_pending_by_height.get(height, 0)
-                    
-                    # Calculate percentage based on type
-                    if percentage_type in ['success', 'failure']:
-                        # For success and failure, use only (success + failure) as denominator
-                        success_failure_total = success_at_height + failure_at_height
-                        if success_failure_total > 0:
-                            if percentage_type == 'success':
-                                percentage = (success_at_height / success_failure_total) * 100
-                            elif percentage_type == 'failure':
-                                percentage = (failure_at_height / success_failure_total) * 100
-                            else:
-                                percentage = 0
-                        else:
-                            percentage = 0
-                    else:
-                        # For pending, use (success + pending + failure) as denominator
-                        total = success_at_height + pending_at_height + failure_at_height
-                        if total > 0:
-                            percentage = (pending_at_height / total) * 100
-                        else:
-                            percentage = 0
-                    
-                    percentages.append(percentage)
-                
-                # Create title and filename based on transaction type
-                if transaction_type == 'cat':
-                    if percentage_type in ['success', 'failure']:
-                        title = f'CAT {percentage_type.title()} Percentage (of Success+Failure) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    else:
-                        title = f'CAT {percentage_type.title()} Percentage (of Success+Pending+Failure) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    filename = f'tx_{percentage_type}_cat_percentage_cutoff.png'
-                elif transaction_type == 'cat_pending_resolving':
-                    title = f'CAT Pending Resolving Percentage (of Resolving+Postponed) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    filename = f'tx_pending_cat_resolving_percentage_cutoff.png'
-                elif transaction_type == 'cat_pending_postponed':
-                    title = f'CAT Pending Postponed Percentage (of Resolving+Postponed) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    filename = f'tx_pending_cat_postponed_percentage_cutoff.png'
-                elif transaction_type == 'regular':
-                    if percentage_type in ['success', 'failure']:
-                        title = f'Regular {percentage_type.title()} Percentage (of Success+Failure) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    else:
-                        title = f'Regular {percentage_type.title()} Percentage (of Success+Pending+Failure) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    filename = f'tx_{percentage_type}_regular_percentage_cutoff.png'
-                else:
-                    if percentage_type in ['success', 'failure']:
-                        title = f'{transaction_type.title()} {percentage_type.title()} Percentage (of Success+Failure) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    else:
-                        title = f'{transaction_type.title()} {percentage_type.title()} Percentage (of Success+Pending+Failure) Over Time - {create_sweep_title(param_name, sweep_type)} (Cutoff Applied)'
-                    filename = f'tx_{percentage_type}_{transaction_type}_percentage_cutoff.png'
-                
-                # Trim the last 10% of data to avoid edge effects
-                if len(heights) > 10:
-                    trim_index = int(len(heights) * 0.9)
-                    heights = heights[:trim_index]
-                    percentages = percentages[:trim_index]
-                
-                if not heights:
-                    continue
-                
-                # Update maximum height
-                if heights:
-                    max_height = max(max_height, max(heights))
-                
-                # Plot with color based on parameter
-                from plot_utils_percentage import create_parameter_label
-                label = create_parameter_label(param_name, param_value)
-                plt.plot(heights, percentages, color=colors[i], alpha=0.7, 
-                        label=label, linewidth=1.5)
-            
-            # Set x-axis limits before finalizing the plot
-            plt.xlim(left=0, right=max_height)
-            
-            plt.title(title)
-            plt.xlabel('Block Height')
-            plt.ylabel(f'{percentage_type.title()} Percentage (%)')
-            plt.grid(True, alpha=0.3)
-            plt.legend(loc="upper right")
-            plt.tight_layout()
-            
-            # Create tx_count_cutoff directory and save plot
-            tx_count_cutoff_dir = f'{results_dir}/figs/tx_count_cutoff'
-            os.makedirs(tx_count_cutoff_dir, exist_ok=True)
-            plt.savefig(f'{tx_count_cutoff_dir}/{filename}', 
-                       dpi=300, bbox_inches='tight')
-            plt.close()
-            
-        except Exception as e:
-            print(f"Warning: Error creating {transaction_type} {percentage_type} percentage cutoff plot: {e}")
-            return
-    
-    # Call the modified function
-    plot_transaction_percentage_cutoff_save(cutoff_data, param_name, results_dir, sweep_type, transaction_type, percentage_type)
